Only print premium warning on command sender site
