+++ velocity/src/main/java/com/github/games647/fastlogin/velocity/listener/ConnectListener.java
@@ -38,7 +38,6 @@ import com.github.games647.fastlogin.velocity.task.FloodgateAuthTask;
 import com.github.games647.fastlogin.velocity.task.ForceLoginTask;
 import com.velocitypowered.api.event.EventTask;
 import com.velocitypowered.api.event.Subscribe;
-import com.velocitypowered.api.event.connection.DisconnectEvent;
 import com.velocitypowered.api.event.connection.PreLoginEvent;
 import com.velocitypowered.api.event.connection.PreLoginEvent.PreLoginComponentResult;
 import com.velocitypowered.api.event.player.GameProfileRequestEvent;
@@ -169,11 +168,5 @@ public class ConnectListener {
 
         // Delay at least one second, otherwise the login command can be missed
         plugin.getScheduler().runAsyncDelayed(loginTask, Duration.ofSeconds(1));
-    }
-
-    @Subscribe
-    public void onDisconnect(DisconnectEvent disconnectEvent) {
-        Player player = disconnectEvent.getPlayer();
-        plugin.getCore().getPendingConfirms().remove(player.getUniqueId());
     }
 }
